Fix traceback in install_wheel() when --always-copy used
The embedded python script in this function imports tempfile, which in
turn needs to import _io.so. The munged pythonpath used for this script
does not contain this directory, causing the embedded script to fail to
import tempfile.

This commit resolves that traceback by adding the lib-dynload dirs from
sys.path to the munged pythonpath.